### 2025-01-17 #7682 @editorialbot
[PRE REVIEW]: IAMReX: an adaptive framework for the multiphase flow and fluid-particle interaction problems
**Submitting author:** @ruohai0925 (<a href="http://orcid.org/0009-0001-7944-3597">Yadong Zeng</a>)
**Repository:** https://github.com/ruohai0925/IAMReX.git
**Branch with paper.md** (empty if default branch): development
**Version:** v1.0.0
