### 2025-01-17 #7682 @arfon
@kyleniemeyer – noting that this is a resubmission of https://github.com/openjournals/joss-reviews/issues/7582
