### 2025-01-17 #7682 @ruohai0925
> [@kyleniemeyer](https://github.com/kyleniemeyer) – noting that this is a resubmission of [#7582](https://github.com/openjournals/joss-reviews/issues/7582)

Hi @arfon @kyleniemeyer, thanks for taking care of that. For potential reviewers, I would recommend the GitHub user YqWangcfd, FR13ndSDP, Jannis Teunissen, or rafmudaf. They have been working on the AMReX based app for several years. Let me know if that works.
