### 2025-01-17 #7682 @ruohai0925
> **Paper file info**:
>
> 📄 Wordcount for `paper.md` is **829**
>
> 🔴 Failed to discover a `Statement of need` section in paper

How to handle this failed to discover a "Statement of need" issue. I think this issue also appeared at the first submission. After that, I added some paragraphs in the paper to emphasize the statement of need about our work.
