### 2025-01-17 #7682 @kyleniemeyer
@ruohai0925 JOSS requires an explicit section with the name "Statement of Need" - so please do add that to your paper (and perhaps move those paragraphs into it)
