### 2025-01-17 #7682 @kyleniemeyer
OK, thanks @ruohai0925. Unfortunately, we do not have an editor available to handle this right now, so I have to put it on our waitlist until someone becomes free.

In the meantime, one thing to consider: I see a `Test` directory, but it appears to only mention a suite of regression tests for IAMR, and not IAMReX. JOSS does require that software submissions have a test suite, so this is something that reviewers will be looking for (and IAMReX would need distinct tests, even if they build on what was sufficient for IAMR, I assume.)
