### 2025-02-15 #7682 @ruohai0925
@kyleniemeyer Hi, Kyle, since our paper has been on the waitlist for more than one month, will there be some available associate editors to help handle that? I can also help find independent researchers to review it if needed.

We also updated the CI/CD test suite specifically for IAMReXs as requested.
