### 2025-04-09 #7682 @piyueh
Hi @esclapez would you be willing to help review this submission for JOSS?
