### 2025-04-14 #7682 @timokoch
@ruohai0925 Two more comments when scanning through the repo and paper: The code is based on IAMR (https://github.com/AMReX-Fluids/IAMR), and by based, I mean it seems to be practically a fork (?) or at least contains literal copies of many files. This raises some questions: (1) why a new repo instead of basing on the existing code? (2) Is the license info correctly stated: the IAMR licence seems to requires explicit mentioning of the license (3) some files like RELEASE_NOTES (last entry from 2022) are also literal copies from IAMR and don't seem to fit the new repo (4) which part of the code is new / different from IAMR and original?

To the last point I know you mention some of this in the paper already. But you have a 2024 publication (https://doi.org/10.1063/5.0236509) about IAMReX (cited in the JOSS paper draft) specifically and I only found (also in the JOSS paper draft) that since then 60 commits 3000 locs have been added. This does not sound so much and it also looks like some of that code was actually cherry-picked from related AMReX repos. At least you should explain what significant improvements have been made to meet the substantial scholarly effort requirement of JOSS.
