### 2025-04-15 #7682 @piyueh
Thank you for the quick response, @timokoch! Forks with this level of differences from the original code are generally acceptable within JOSS. If the current draft doesn’t yet clearly explain those differences and this is raising concerns, the authors are expected to work with reviewers, refine the submission, and clarify concerns.

While the review process hasn’t formally started yet—we’re still in the process of finding a second reviewer—it might be helpful for @ruohai0925 to begin considering this feedback. That way, once the review begins, things may move along more smoothly.
