### 2025-04-16 #7682 @piyueh
Hi @ruohai0925 I noticed that in the IAMReX repo, there are several commits that seem to do a manual copy of source files from the upstream ([AMReX-Fluids/IAMR](https://github.com/AMReX-Fluids/IAMR)). For example:

* https://github.com/ruohai0925/IAMReX/commit/2eba74f83de899883ab0540b2569a1ac871d2935
* https://github.com/ruohai0925/IAMReX/commit/7e7edcbbaabd1aa34b8b397d40aad217960426d7
* https://github.com/ruohai0925/IAMReX/commit/122d89c4da7003d880846bef3f50bf653e778655
* https://github.com/ruohai0925/IAMReX/commit/c4ca2e2bb175a7248e230bde491d62148cb32158

They correspond to the [PRs](https://github.com/AMReX-Fluids/IAMR/pulls?q=is%3Apr+is%3Aclosed) in the upstream repo. JOSS does not solely rely on the commit history to verify substantial scholarly effort. So this may be fine. But by hard copying, it may make reviewers hard to identify the actual work done by this specific submission. So it would be great if you can address where new code or modifications are in this work, compared to the upstream code. And thus reviewers can have a better idea where to look into for your work and what belong to the upstream.
