### 2025-04-17 #7682 @smiths
@piyueh, yes, I would be interested in reviewing the submission.  I haven't done a JOSS review previously, so the checklist you mentioned would be very helpful.  :smile:
