### 2025-04-18 #7682 @piyueh
Hi @smiths — thanks for your interest! I found your info in our reviewer database and think your expertise would be a great fit for this submission.

Here’s what to expect: our bot (`@editorialbot`) will soon add you as a reviewer and open a new GitHub issue for the actual review. (The current issue here is just for assigning reviewers.) Once that’s set up, you’ll get a checklist to guide the process — here’s an example of the checklist: https://github.com/openjournals/joss-reviews/issues/7895#issuecomment-2725210688

The review is iterative: reviewers raise issues as needed, and authors respond or update their code. Our goal is to help the submission meet JOSS standards, not just evaluate it. If you're interested in more details, here's a full documentation of the JOSS review process: https://joss.readthedocs.io/

Feel free to ping me if you have any questions!
