### 2025-04-24 #7682 @ruohai0925
> Hi [@ruohai0925](https://github.com/ruohai0925) I noticed that in the IAMReX repo, there are several commits that seem to do a manual copy of source files from the upstream ([AMReX-Fluids/IAMR](https://github.com/AMReX-Fluids/IAMR)). For example:
>
> * [ruohai0925/IAMReX@2eba74f](https://github.com/ruohai0925/IAMReX/commit/2eba74f83de899883ab0540b2569a1ac871d2935)
> * [ruohai0925/IAMReX@7e7edcb](https://github.com/ruohai0925/IAMReX/commit/7e7edcbbaabd1aa34b8b397d40aad217960426d7)
> * [ruohai0925/IAMReX@122d89c](https://github.com/ruohai0925/IAMReX/commit/122d89c4da7003d880846bef3f50bf653e778655)
> * [ruohai0925/IAMReX@c4ca2e2](https://github.com/ruohai0925/IAMReX/commit/c4ca2e2bb175a7248e230bde491d62148cb32158)
>
> They correspond to the [PRs](https://github.com/AMReX-Fluids/IAMR/pulls?q=is%3Apr+is%3Aclosed) in the upstream repo. JOSS does not solely rely on the commit history to verify substantial scholarly effort. So this may be fine. But by hard copying, it may make reviewers hard to identify the actual work done by this specific submission. So it would be great if you can address where new code or modifications are in this work, compared to the upstream code. And thus reviewers can have a better idea where to look into for your work and what belong to the upstream.

Hi @piyueh, you are right! Some of the commits are inherited from the original IAMR since they fix some minor issues which are not related to our own scholar contributions. To highlight the new code or modifications, one way, maybe, is for us to clone a clean original IAMR repo, ignore its commit histories, and condense our own contributions into the new commits.

We were thinking about doing this before, yet it will make the commit history and age of software be very short and users/reviewers might think we have not spent too much working time on this. Actually, we have started working on it since 2022 based on IAMR.

But if you think it is eaiser for the users to catch up with what we have done, maybe it is a good time for us to isolate the old commits, and split our own contributions into new commits.

@piyueh @timokoch , thanks for your efforts! Let us continue our discussion in the new PR.
