+++ Source/NavierStokesBase.cpp
@@ -27,6 +27,7 @@
 #include <AMReX_EB_Redistribution.H>
 #include <AMReX_EBMultiFabUtil_C.H>
 #endif
+
 #ifdef AMREX_USE_TURBULENT_FORCING
 #include <TurbulentForcing_params.H>
 #endif
