### 2025-05-18 #8080 @ruohai0925
@piyueh Reagrding the contribution and authorship question by @smiths and @timokoch, I have contributed 101 commits with the author name yzeng and another 35 commints with the author name ruohai0925. For all other authors, they have contributed some commits for different parts of IAMeX codes.

As clarified in the paper draft, some parts of the IAMReX codes inherit from the [IAMR](https://github.com/AMReX-Fluids/IAMR) repo. then I sent an email to major contributors of IAMR repo last week. Here are our [conversations](https://drive.google.com/file/d/1Oca0rdyOytsZWkN2oh5byAB86YyZ3Xzt/view?usp=sharing). From their replies, most major contributors of IAMR are happy to appear in the acknowledgements but do not want to be a co-author.

In addition, with the help from Timo, the community guidelines also have been updated. More updates are on the way.
