### 2025-05-26 #8080 @ruohai0925
Hi Timo,

Regarding your questions.

a). Yes, IAMR also has GPU support and sub-cycling scheme, yet they only focus on the sinlge-phase incompressible flow. Our work IAMReX extends its ability to include, 1) multiphase flow using level set method and reinitializaiton technique; 2) diffused immersed boundary method (DIBM) for fluid structure interaction problems with particle-collision models; We emphasize these contributions in the paper, and these new contributions, of course, need to support the running on GPUs and can also utilize the sub-cycling scheme and non-subcycling scheme;

b). The last POF paper focuses more on the validation cases of the DIBM and it application on the particle-laden flow. It does not disclose the implementation details of the DIBM method in the IAMReX codes, and that is why we think an additional JOSS draft might be a good plus. We also add the level set method, reinitialization technique, and some two-phase flow validation cases in the IAMReX. Finally, we summarized and added a bunch of cases to help us better evaluate how AMR-related parameters influence the efficiency.

I can point out the implementation of these codes if needed.

Thanks!
