### 2025-06-08 #8080 @ruohai0925
@piyueh @timokoch Regarding the lines changed compared with the original IAMR code, we made a summary and prepared the following table for reference, in case it is needed for code review. The above scholar efforts are included in the lines of following files. Maybe it is better for us to put the info of these added/modified functions in the documentation. Let me what you think.

![Image](https://github.com/user-attachments/assets/fb9d49ea-fde1-4681-bfd4-9b41e18fa43a)
