### 2025-06-10 #8080 @timokoch
> I can point out the implementation of these codes and/or add this info into the draft if needed.

Yes please document this. Best both in the code documentation and the paper.
