### 2025-06-13 ruohai0925/IAMReX#18 @smiths
README Install Instructions Do Not Match Web-page Instructions
Step 3 of the [Install Instructions in the README file ](https://github.com/ruohai0925/IAMReX/blob/development/README.md#Install) do not match the Step 3 instructions in the [Getting Started document](https://ruohai0925.github.io/IAMReX/Getting_Started.html).  The README instructions incorrectly include extra information in the name of the repo (`tree/development`).

This minor issue points to a larger problem.  Very similar information is maintained in two different places.  Can you put the installation instructions in one spot and link to them where you need them?
