+++ docs/_sources/Software_Chapter.rst.txt
@@ -1,0 +1,6 @@
+.. role:: cpp(code)
+   :language: c++
+
+
+Scholarly effort of IAMReX
+==========================
