+++ docs/_sources/DiffusedIBM.rst.txt
@@ -1,0 +1,11 @@
+
+Direct-Forcing Immersed Boundary Method
+=======================================
+
+
+Time advancement
+----------------
+
+
+Types of kinematic constraints
+------------------------------
