+++ docs/_sources/Algorithm_Chapter.rst.txt
@@ -1,0 +1,13 @@
+.. _Chap:Algorithm:
+
+
+
+Algorithm of IAMReX
+===================
+
+.. toctree::
+
+    FluidEquations
+    LevelSet
+    DiffusedIBM
+    IAMRandIAMReX
