+++ docs/_sources/Algorithm.rst.txt
@@ -1,0 +1,9 @@
+
+Algorithm and Software Details
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. toctree::
+   :caption: Contents:
+
+   Algorithm_Chapter
+   Software_Chapter
