### 2025-06-20 ruohai0925/IAMReX#18 @ruohai0925
This issue was fixed in this PR: https://github.com/ruohai0925/IAMReX/pull/19. We updated the README.md and put the installation instruction in the documentation.
