+++ docs/_sources/Software_Chapter.rst.txt
@@ -1,6 +1,203 @@
-.. role:: cpp(code)
-   :language: c++
-
 
 Scholarly effort of IAMReX
 ==========================
+
+IAMeX represents a significant enhancement to the baseline IAMR (Incompressible Adaptive Mesh Refinement) code, introducing advanced multi-physics capabilities for complex multi-phase flow and fluid-solid interaction simulations. This analysis identifies and summarizes the key innovations and contributions that distinguish IAMeX from the standard IAMR implementation.
+
+I. Two Enhanced Time Advancement Methods
+----------------------------------------
+
+New Specialized Advance Functions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+IAMeX introduces two distinct time advancement methods, each targeting specific multi-physics scenarios:
+
+1. Two-Phase Flow with Level Set Method
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+- **Function**: ``NavierStokes::advance_semistaggered_twophase_ls()``
+- **Purpose**: Sharp interface tracking for immiscible two-phase flows
+- **Key Features**:
+  - Level set field evolution with both non-conservative and convervative advection schemes
+  - Level-by-level reinitialization functions for conserving mass and avoid shape distortion
+  - Interface-dependent density and viscosity updates
+  - Heaviside function computation for material properties
+
+2. Fluid-Structure Interaction with Diffused Immersed Boundary Method (DIBM) and Particle Collision
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+- **Function**: ``NavierStokes::advance_semistaggered_fsi_diffusedib()``
+- **Purpose**: Robust fluid-solid coupling without mesh conformity
+- **Key Features**:
+  - Particle-based immersed boundary implementation
+  - 6-DOF rigid body dynamics
+  - Collision detection and resolution
+
+II. Level Set Interface Tracking
+--------------------------------
+
+Enhanced State Management
+^^^^^^^^^^^^^^^^^^^^^^^^^
+- **Extended State Vector**: ``NUM_STATE_MAX`` increased from 4 to 5 components
+- **Level Set Component**: ``phicomp`` for interface representation
+- **Signed Distance**: Maintained through reinitialization
+
+Key Computational Functions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+- ``NavierStokesBase::get_phi_half_time()``: Get the mid point Level set function
+- ``NavierStokesBase::phi_to_heavi()``: Level set to Heaviside function conversion
+- ``NavierStokesBase::phi_to_sgn0()``: Level set to Sign function conversion
+- ``NavierStokesBase::heavi_to_rhoormu()``: Material property mapping
+- ``NavierStokesBase::reinit()``: Distance function maintenance using reinitialization
+- ``NavierStokesBase::reinitialization_sussman()``: Reinitialization with Sussman's method
+- ``NavierStokesBase::reinitialization_consls``: Reinitialization for the conservative LS method
+- ``NavierStokesBase::rk_first_reinit()``: First RK step during reinitialization
+- ``NavierStokesBase::rk_second_reinit()``: Second RK step during reinitialization
+- ``NavierStokesBase::mass_fix()``: Mass fix subroutine during reinitialization
+
+Level Set Methodology
+^^^^^^^^^^^^^^^^^^^^^
+- **Sharp Interface**: Maintains interface thickness of 1.5~2 grid cells
[…]
