+++ docs/_sources/Results.rst.txt
@@ -3,6 +3,108 @@
 Validation Cases
 ================
 
+.. _R-PVF:
+
+Particle Volume Fraction (PVF)
+------------------------------
+
+The particle volume fraction (PVF), which joins the calculation of free motion updates, is introduced and validated in this session. The PVF is approximated by the signed-distance level-set function :math:`\phi` of the fluid-particle interface. The level-set function :math:`\phi` is located at cell nodes and is calculated at the eight corners of each cell. The symbol of :math:`\phi` as well as the intersected interface is shown in :ref:`figPvfLevelSet`. Here, :math:`\phi` is negative inside the particle and positive outside the particle.
+
+.. _figPvfLevelSet:
+
+.. figure:: ./Results/pvf_level_set_function.png
+    :align: center
+    :width: 5.0in
+    :alt: Level-set function at grid cell corners
+
+    Sketch of :math:`\phi` at each corner of a grid cell with interface.
+
+Based on the level-set function :math:`\phi`, an approximation of PVF can be obtained by the following equation:
+
+.. math::
+
+   \alpha_{i,j,k} = \frac{\sum_{m=1}^{8}-\phi_mH(-\phi_m)}{\sum_{m=1}^8|\phi_m|},
+
+where :math:`H` is the Heaviside function, defined by:
+
+.. math::
+
+   H(\phi) = 
+   \begin{cases}
+   0, & \phi \le 0\\
+   1, & \phi > 0
+   \end{cases}
+
+On the right-hand side of equation :math:`\alpha_{i,j,k}`, the value of :math:`\phi` for each cell corner depends on the location of the fluid-particle interface. When the shape of the particle surface is analytically given, the :math:`\phi` value can be determined by calculating the Euclidean distance from the corner point of the cell to the particle surface. The calculation of PVF is then transformed from an exact integral to a numerical integral. As shown in :ref:`figPvfCase`, the cell value of PVF varies from :math:`0` to :math:`1`, depending on the relative position between the cell center and the fluid-particle interface.
+
+.. _figPvfCase:
+
+.. figure:: ./Results/pvf_cases_of_exact_value.png
+    :align: center
+    :width: 5.0in
+    :alt: PVF cases with different values
+
+    Sketch of PVF with different values. The white and green colors refer to the cell areas within the fluid and particle, respectively.
+
+We validate the correctness and convergence of the above PVF approximation by calculating the volumes of spherical and ellipsoidal surfaces on the Cartesian grid. The exact solutions for the sphere and ellipsoid are given by:
+
+.. math::
+   :label: equ:exact solutions of the sphere and ellipsoid
+
+   V_{exact} = 4\pi \frac{abc}{3}
+
+where :math:`a`, :math:`b`, and :math:`c` are the semi-axes of the ellipsoid. For a sphere case, we have :math:`a=b=c`. The signed-distance level-set function is:
+
+.. math::
+   :label: equ:level-set function of spheroidal geometries
+
+   \phi_{i,j,k}= \sqrt{\frac{(x_{i,j,k}-x_{p})^2}{a^2} + \frac{(y_{i,j,k}-y_{p})^2}{b^2} + \frac{(z_{i,j,k}-z_{p})^2}{c^2}}-1,
+
[…]

[…]
