+++ docs/_sources/DiffusedIBM.rst.txt
@@ -36,7 +36,7 @@ where the convective term :math:`{\nabla} \cdot \left(\mathbf{u}\mathbf{u} \righ
     :align: center
     :width: 5.0in
 
-    The number of markers change with the :math:`d/h` for a single particle scenario, where ::math:`d` is the particle diameter and ::math:`h` is the grid spacing.
+    The number of markers change with the :math:`d/h` for a single particle scenario, where :math:`d` is the particle diameter and :math:`h` is the grid spacing.
 
 **Step 2**: The updated velocity :math:`\widetilde{\mathbf{u}}^{*, n+1}` needs to be corrected to satisfy the no-slip boundary condition at the fluid–particle interfaces :math:`\partial V_{\mathrm{b}}(t)`. This step is divided into four substeps in Algorithm-1_. We first interpolate the intermediate Eulerian Velocity obtained from **Step 1** to the Lagrangian Velocity of markers. The Lagrangian forces are then calculated based on the desired velocity at the interface and the intermediate velocity. Next, the Eulerian forces are obtained from the spreading of Lagrangian forces by using either the three-point or four-point delta function. As shown in :ref:`figinterpolation`, the Eulerian cells, enclosed by red dashed circles, refer to grid areas that are influenced by two blue markers. These two Lagrangian markers also share some intersected areas, which are marked by green arrows. Finally, the Eulerian velocity is corrected by the updated Eulerian Force.
 
@@ -141,3 +141,5 @@ In contrast to the prescribed kinematics case, the motion of a freely moving par
 On the right-hand side of :eq:`eq:particle1`-:eq:`eq:particle2`, the term :math:`\mathbf{F}_l^{n+1/2}` refers to the Lagrangian Force, coming from the final value of :math:`\mathbf{F}^{m}` in Algorithm-1_. The time derivatives of momentum integration :math:`\rho_f \frac{d}{d t}\left(\int_{V_p} \mathbf{u} d V\right)` and angular momentum integration :math:`\rho_f \frac{d}{d t}\left(\int_{V_p} \mathbf{r} \times \mathbf{u} d V\right)` within the particle are also included. These two integrated terms account for flow unsteadiness by using the PVF field (:ref:`R-PVF`). The term :math:`\left(\rho_p-\rho_f\right) V_p \mathbf{g}` considers the buoyancy effects. The terms :math:`\mathbf{F}_c^{n+1 / 2}` and :math:`\mathbf{T}_c^{n+1 / 2}` refer to the induced force and torque generated by the particle collision, respectively. If there is only one single particle in the system, both :math:`\mathbf{F}_c^{n+1 / 2}` and :math:`\mathbf{T}_c^{n+1 / 2}` are set to be zero. On the left-hand side of :eq:`eq:particle1`-:eq:`eq:particle2`, we use the second-order mid-point scheme to integrate particle motions . After updating the particle centroid velocity :math:`\mathbf{U}_{r}^{n+1}` and angular velocity :math:`\mathbf{W}_{r}^{n+1}` at :math:`t^{n+1}`, we go back to :eq:`eq:pr2` to update new position of the particle centroid :math:`\mathbf{X}_{r}^{n+1}`.
 
[…]

[…]
