### 2025-06-20 #8080 @ruohai0925
Some updates:

1. This issue is fixed: https://github.com/ruohai0925/IAMReX/issues/18.

2. We added a Scholarly effort session in the documentation： https://ruohai0925.github.io/IAMReX/Software_Chapter.html. I will also try to include the key points in the draft later.

Keep tuning and improving the documentation.
