+++ docs/_sources/LevelSet.rst.txt
@@ -1,3 +1,12 @@
 
-Level Set
+Level Set Method
 =========
+
+The level set method is a powerful tool for capturing the interface between two immiscible fluids. It is based on the idea of a signed distance function, which is a function that assigns a signed distance to each point in the domain. The level set function is defined as follows:
+
+.. math::
+
+    \phi(\mathbf{x}, t) = 0
+
+where :math:`\phi` is the level set function, :math:`\mathbf{x}` is the position vector, and :math:`t` is the time.
+
