+++ docs/_sources/LevelSet.rst.txt
@@ -1,6 +1,7 @@
+.. _LevelSetMethod:
 
 Level Set Method
-=========
+===============
 
 The level set method is a powerful tool for capturing the interface between two immiscible fluids. It is based on the idea of a signed distance function, which is a function that assigns a signed distance to each point in the domain. The level set function is defined as follows:
 
@@ -9,4 +10,121 @@ The level set method is a powerful tool for capturing the interface between two 
     \phi(\mathbf{x}, t) = 0
 
 where :math:`\phi` is the level set function, :math:`\mathbf{x}` is the position vector, and :math:`t` is the time.
+
+Material Properties
+------------------
+
+The density and viscosity are defined as functions of the level set field:
+
+.. math::
+   :label: eq:rho
+
+   \rho(\phi) = \rho_1 H(\phi) + \rho_2 (1 - H(\phi))
+
+.. math::
+   :label: eq:mu
+
+   \mu(\phi) = \mu_1 H(\phi) + \mu_2 (1 - H(\phi))
+
+where :math:`H(\phi)` is the Heaviside function:
+
+.. math::
+
+   H(\phi) = \begin{cases}
+   1, & \phi > 0 \\
+   0, & \phi \leq 0
+   \end{cases}
+
+Time Discretization
+------------------
+
+For a single level, the momentum equation :eq:`eq:ns` is advanced by a fractional step method with the approximate projection :cite:`rider1995approximate,rider1998filtering,almgren2000approximate` to enforce the incompressibility condition (equation :eq:`eq:div`). The LS advection equation :eq:`eq:phi` is updated using the Godunov scheme :cite:`colella1990multidimensional,bell1989second,almgren1998conservative,sussman1999adaptive`.
+
+At the beginning of each time advancement of level :math:`l`, the velocity :math:`\boldsymbol{u}^{n,l}` and the LS function :math:`\phi^{n,l}` at time :math:`t^{n,l}` are given. Owing to the fractional step method we used, the pressure is staggered in time :cite:`almgren1998conservative,sussman1999adaptive,bhalla2013unified` and thus the pressure :math:`p^{n-1/2,l}` at time :math:`t^{n-1/2,l}` is known. Because the single-level advancement concerns only level :math:`l`, we omit the superscript :math:`{}^l` in this section. To obtain the updated velocity :math:`\boldsymbol{u}^{n+1}`, pressure :math:`p^{n+1/2}`, and LS function :math:`\phi^{n+1}` on level :math:`l`, the solver performs the following steps:
+
+1. **Advance the LS function** as
+
+   .. math::
+      :label: eq:s0phin1
+
+      \phi^{n+1} = \phi^{n} - \Delta t \left[\nabla \cdot (\boldsymbol{u}\phi) \right]^{n+1/2}
+
+   The advection term in the above equation is calculated using the Godunov scheme.
+
+2. **Solve the intermediate velocity** :math:`\boldsymbol{u}^{*}` semi-implicitly
+
+   .. math::
+      :label: eq:viscsolve
+
+      \begin{aligned} 
+      &\boldsymbol{u^{*,n+1}}-\frac{\Delta t}{2\rho(\phi^{n+1/2})Re}{\nabla} \cdot {\mu(\phi^{n+1})}{\nabla}\boldsymbol{u^{*,n+1}} = 
+      \boldsymbol{u^n}-\Delta t \left[\nabla \cdot (\boldsymbol{uu})\right]^{n+1/2} + \\ &\frac{\Delta t}{\rho(\phi^{n+1/2})}\bigg[-\nabla p^{n-1/2}+ 
[…]
