+++ docs/_sources/LevelSet.rst.txt
@@ -38,9 +38,9 @@ where :math:`H(\phi)` is the Heaviside function:
 Time Discretization
 ------------------
 
-For a single level, the momentum equation :eq:`eq:ns` is advanced by a fractional step method with the approximate projection :cite:`rider1995approximate,rider1998filtering,almgren2000approximate` to enforce the incompressibility condition (equation :eq:`eq:div`). The LS advection equation :eq:`eq:phi` is updated using the Godunov scheme :cite:`colella1990multidimensional,bell1989second,almgren1998conservative,sussman1999adaptive`.
+For a single level, the momentum equation :eq:`eq:ns` is advanced by a fractional step method with the approximate projection to enforce the incompressibility condition (equation :eq:`eq:div`). The LS advection equation :eq:`eq:phi` is updated using the Godunov scheme.
 
-At the beginning of each time advancement of level :math:`l`, the velocity :math:`\boldsymbol{u}^{n,l}` and the LS function :math:`\phi^{n,l}` at time :math:`t^{n,l}` are given. Owing to the fractional step method we used, the pressure is staggered in time :cite:`almgren1998conservative,sussman1999adaptive,bhalla2013unified` and thus the pressure :math:`p^{n-1/2,l}` at time :math:`t^{n-1/2,l}` is known. Because the single-level advancement concerns only level :math:`l`, we omit the superscript :math:`{}^l` in this section. To obtain the updated velocity :math:`\boldsymbol{u}^{n+1}`, pressure :math:`p^{n+1/2}`, and LS function :math:`\phi^{n+1}` on level :math:`l`, the solver performs the following steps:
+At the beginning of each time advancement of level :math:`l`, the velocity :math:`\boldsymbol{u}^{n,l}` and the LS function :math:`\phi^{n,l}` at time :math:`t^{n,l}` are given. Owing to the fractional step method we used, the pressure is staggered in time and thus the pressure :math:`p^{n-1/2,l}` at time :math:`t^{n-1/2,l}` is known. Because the single-level advancement concerns only level :math:`l`, we omit the superscript :math:`{}^l` in this section. To obtain the updated velocity :math:`\boldsymbol{u}^{n+1}`, pressure :math:`p^{n+1/2}`, and LS function :math:`\phi^{n+1}` on level :math:`l`, the solver performs the following steps:
 
 1. **Advance the LS function** as
 
@@ -92,7 +92,7 @@ At the beginning of each time advancement of level :math:`l`, the velocity :math
 
       \boldsymbol{{u^{n+1}}} = \Delta t \left(\boldsymbol{V} - \frac{1}{\rho^{n+1/2}} \nabla p^{n+1/2}\right)
 
[…]
