+++ docs/_sources/Algorithm.rst.txt
@@ -6,4 +6,3 @@ Algorithm and Software Details
 
    Algorithm_Chapter
    Software_Chapter
-   LevelSet
