### 2025-06-20 #8080 @ruohai0925
> [@piyueh](https://github.com/piyueh) [@timokoch](https://github.com/timokoch) Regarding the lines changed compared with the original IAMR code, we made a summary and prepared the following table for reference, in case it is needed for code review. The above scholar efforts are included in the lines of following files. Maybe it is better for us to put the info of these added/modified functions in the documentation. Let me what you think.
>
> ![Image](https://github.com/user-attachments/assets/46cb5a93-f770-4530-828a-a6e267fae2b8)

And we updated the table here to include the final line contributions of our code repo.
