### 2025-06-25 #8080 @ruohai0925
Hi @piyueh, thanks for the detailed comments! I will update the paper draft carefully to deal with each point you mentioned, including modifying the paragraphs to avoid the wrong impression, adding paragraphs to highlight the scolarly efforts, adding explanations to clarify how the latest IAMReX code connects with the our previous publications and what are the new developements.

Please give me some time. At the same time. if @smiths and @timokoch have more concerns, please also let me know.

Thanks!
