+++ docs/_sources/DiffusedIBM.rst.txt
@@ -18,7 +18,7 @@ Time advancement
 
     Sketch of the two-dimensional semi-staggered grid and variable locations. The blue triangles, red squares, and black circles represent cell-centered variables, node-centered variables, and interface Lagrangian markers.
 
-To solve the partial differential equations of :ref:`FluidEquationsPart`, the canonical projection is 
+To solve the partial differential equations of :ref:`FluidEquationsPart`, the canonical projection :cite:`chorin1967numerical, almgren1998conservative` is
 applied to the semi-staggered grid. As shown in :numref:`figSemiStaggeredGrid`, the fluid velocity (:math:`u` and :math:`v`), the Eulerian force :math:`f`, and particle volume fraction :math:`\alpha` are located at the cell center. The pressure :math:`p` and level set function :math:`\phi` are at the node center. The temporal and spatial discretizations of equations for single-level advancement are considered here. At the time :math:`t^{n}`, the Eulerian velocity :math:`\mathbf{u}^{n}` and pressure :math:`p^{n-1/2}` are known. The particle position :math:`\mathbf{X}^{n}_l` and velocity :math:`\mathbf{U}\left(\mathbf{X}^{n}_l\right)` are also available. The time advancement during the interval :math:`[t^{n}, t^{n+1}]` proceeds as follows.
 
 
@@ -29,7 +29,7 @@ applied to the semi-staggered grid. As shown in :numref:`figSemiStaggeredGrid`, 
 
     \rho_{f}\left(\frac{\widetilde{\mathbf{u}}^{*,n+1}-\mathbf{u}^{n}}{\Delta t}+\nabla \cdot \left(\mathbf{u}\mathbf{u} \right)^{n+\frac{1}{2}}\right)=-\nabla p^{n-\frac{1}{2}}+\frac{1}{2}\left(\nabla \cdot \mu \nabla{\widetilde{\mathbf{u}}^{*,n+1}}+{\nabla} \cdot \mu \nabla {\mathbf{u}}^{n}\right)+\rho_{f}\mathbf{g}
 
-where the convective term :math:`{\nabla} \cdot \left(\mathbf{u}\mathbf{u} \right)^{n+\frac{1}{2}}` is calculated using the second-order Godunov scheme. The Godunov algorithm has been proven to be a very robust method for handling conservation laws, especially for high Reynolds number flows. Many researchers who previously worked on incompressible fluid simulations have applied this method to deal with the convection terms in equations. In our previous work, we provided a detailed derivation of the Godunov algorithm. In our recent work, we applied this algorithm to simulate two-phase flows at high Reynolds numbers (i.e., :math:`Re\approx 10^{6}`) and obtained stable and physically meaningful results. In this step, only the pure fluid system is solved and no particle-related influence is included.
[…]
