+++ docs/_sources/references.rst.txt
@@ -4,4 +4,3 @@ References
 
 .. bibliography:: refs.bib
    :style: unsrt
-   :all:
