+++ docs/_sources/LevelSet.rst.txt
@@ -1,7 +1,7 @@
 .. _LevelSetMethod:
 
 Level Set Method
-===============
+================
 
 The level set method is a powerful tool for capturing the interface between two immiscible fluids. It is based on the idea of a signed distance function, which is a function that assigns a signed distance to each point in the domain. The level set function is defined as follows:
 
@@ -12,7 +12,7 @@ The level set method is a powerful tool for capturing the interface between two 
 where :math:`\phi` is the level set function, :math:`\mathbf{x}` is the position vector, and :math:`t` is the time.
 
 Material Properties
-------------------
+-------------------
 
 The density and viscosity are defined as functions of the level set field:
 
@@ -36,7 +36,7 @@ where :math:`H(\phi)` is the Heaviside function:
    \end{cases}
 
 Time Discretization
-------------------
+-------------------
 
 For a single level, the momentum equation :eq:`eq:ns` is advanced by a fractional step method with the approximate projection to enforce the incompressibility condition (equation :eq:`eq:div`). The LS advection equation :eq:`eq:phi` is updated using the Godunov scheme.
 
@@ -56,9 +56,9 @@ At the beginning of each time advancement of level :math:`l`, the velocity :math
    .. math::
       :label: eq:viscsolve
 
-      \begin{aligned} 
-      &\boldsymbol{u^{*,n+1}}-\frac{\Delta t}{2\rho(\phi^{n+1/2})Re}{\nabla} \cdot {\mu(\phi^{n+1})}{\nabla}\boldsymbol{u^{*,n+1}} = 
-      \boldsymbol{u^n}-\Delta t \left[\nabla \cdot (\boldsymbol{uu})\right]^{n+1/2} + \\ &\frac{\Delta t}{\rho(\phi^{n+1/2})}\bigg[-\nabla p^{n-1/2}+ 
+      \begin{aligned}
+      &\boldsymbol{u^{*,n+1}}-\frac{\Delta t}{2\rho(\phi^{n+1/2})Re}{\nabla} \cdot {\mu(\phi^{n+1})}{\nabla}\boldsymbol{u^{*,n+1}} =
+      \boldsymbol{u^n}-\Delta t \left[\nabla \cdot (\boldsymbol{uu})\right]^{n+1/2} + \\ &\frac{\Delta t}{\rho(\phi^{n+1/2})}\bigg[-\nabla p^{n-1/2}+
       \frac{1}{2Re}{\nabla}\cdot{\mu(\phi^{n})}{\nabla}\boldsymbol{u^n} + \rho(\phi^{n+1/2}) \frac{z}{Fr^2} -  \frac{1}{We}\kappa(\phi^{n+1/2})\delta(x^{n+1/2})\boldsymbol{n}\bigg].
       \end{aligned}
 
@@ -74,25 +74,25 @@ At the beginning of each time advancement of level :math:`l`, the velocity :math
 3. **Apply the projection method** to obtain the pressure and a solenoidal velocity field. To conduct the level projection, a temporary variable :math:`\boldsymbol{V}` is defined as
 
    .. math::
-      :label: eq:ns_lp1
+      :label: eq:ns_lp_ls1
 
       \boldsymbol{V} =  \frac{\boldsymbol{{u^{*,n+1}}}}{\Delta t} + \frac{1}{\rho(\phi^{n+1/2})} \nabla p^{n-1/2}
 
    Then the updated pressure :math:`p^{n+1/2}` is calculated by
 
    .. math::
-      :label: eq:ns_lp2
+      :label: eq:ns_lp_ls2
 
[…]
