### 2025-06-27 ruohai0925/IAMReX#23 @smiths
Two License Files Confusion
Clarification should be added for the [two license files](https://github.com/ruohai0925/IAMReX/tree/development/LICENSES).  The original license file for IAMR is fine, but how the new license file fits in should be clarified.  I believe the convention is for the second license file to be named `NOTICE` or `CONTRIBUTING-LICENSE`.  I believe it is also supposed to include a statement that it is a fork of the original project and include a pointer to the original license file.

If possible, it would be nice for GitHub to detect the `LICENSE` and display it in the sidebar.  I believe only one license is detected, but the custom license could be pointed to from the README file.
