### 2025-06-27 ruohai0925/IAMReX#25 @smiths
Provide More Detail for Running the Code
The README file provides instructions on [running the code](https://ruohai0925.github.io/IAMReX/Getting_Started.html#running-the-code), but it is vague on how to use the plot files and the checkpoint files.  Please provide more information to the user.  My guess is that the details on the file format and how to use the checkpoints are already documented somewhere, possibly external to the current repo.  You can just point to that documentation.
