### 2025-06-27 ruohai0925/IAMReX#26 @smiths
Reproducing Examples in Documentation
The [JOSS paper](https://github.com/openjournals/joss-papers/blob/joss.08080/joss.08080/10.21105.joss.08080.pdf) states that input files, profiling scripts and raw post-processing data are available for reproducing all results, but the traceability between the examples in the documentation and their source files is unclear.  For instance, where are the source files for the [sample cases](https://ruohai0925.github.io/IAMReX/Introduction_Chapter.html#sample-cases) in the IAMReX documentation?  Ideally, a link to the source files should be provided in the documentation.  Reproducing these results is part of the requirement for reviewing a JOSS submission.
