### 2025-06-27 ruohai0925/IAMReX#27 @smiths
Figure 2 Conclusion in IAMReX Documentation?
Please add the conclusion that the reader is being asked to make for Figure 2 in the [Sample Cases in the IAMReX Documentation](https://ruohai0925.github.io/IAMReX/Introduction_Chapter.html#sample-cases).  What is the difference between a and b that you are trying to highlight?
