### 2025-06-27 ruohai0925/IAMReX#28 @smiths
Figure 3 References Missing
Figure 3 in the Sample Cases has two citations (Guermond et al and Tryggvason), but no publications by these authors appear in the References.  Reproducibility of these results is not possible without knowing what is being compared.  Also, it is ambiguous to talk about the tip location of the falling fluid and the rising fluid.  What simulation is being referenced here?  Inadequate information is provided for reproducibility.

It feels like these figures were pulled from another document that provided more context.  If that is a case, a pointer to that other document would be helpful.
