### 2025-06-27 ruohai0925/IAMReX#29 @smiths
Figure 4 - inadequate information to reproduce
In the [Sample Cases](https://ruohai0925.github.io/IAMReX/Introduction_Chapter.html#sample-cases), inadequate information is given to reproduce Figure 4.  This is related to #26.
