### 2025-06-27 #8080 @ruohai0925
> Thanks [@ruohai0925](https://github.com/ruohai0925) for updates! Typically, once the reviewers begin their evaluation, I refrain from commenting further to give them space to complete their assessments. However, since the scholarly contribution has become a focal point in this review, I’d like to share a few thoughts. This submission is not typical because it is a fork, meaning upstream source code is also part of the submission and is modified directly in-place, while the upstream developers are not listed as authors. So it's especially important to distinguish the contributions of the authors of this submission from those of the upstream developers.
>
> * For the code size, adding around 3,000 new lines to the core functionality is considered acceptable per the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), especially considering the extra work put into testing, tutorials, and documentation.
> * From what I can tell, the reviewers seem more concerned with how things are worded in the paper and documentation. For example, the paragraph that introduces IAMReX as being built on top of IAMR says:
>   > In addition, IAMReX is a publicly accessible platform designed specifically for developing massively parallel block-structured adaptive mesh refinement (BSAMR) applications. The code now supports hybrid parallelization using either pure MPI or MPI & OpenMP for ...
>
>
>   This might give readers an impression that IAMR doesn't already support BSAMR/MPI/OpenMP, which could make it sound like those features are unique to IAMReX. But based on your reply, IAMR already has those capabilities and IAMReX is building on top of them. So it seems like a wording issue, and adjusting the language could help make that clearer.
> * The JOSS paper and README could expand a bit more on the differences between IAMReX and IAMR, including but not limited to methods, algorithms, and capabilities. I believe in your documentation, [this page](https://ruohai0925.github.io/IAMReX/IAMRandIAMReX.html) and [this one](https://ruohai0925.github.io/IAMReX/Software_Chapter.html) already explain how IAMReX and IAMR are related and where the differences are in the source code. So it might be helpful to add links to them in the README or JOSS paper so reviewers/readers can more easily see the differences.
> * Another point to clarify is what this JOSS submission covers. Is it meant to supplement [your earlier publication](https://doi.org/10.1063/5.0236509) as an artifact? Or is it describing new developments beyond that publication? According to the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software), it is acceptable for a JOSS submission to supplement previous work. If that is the case here, the paper should clearly state what previous work it is supplementing. On the other hand, if this submission adds new developments on top of earlier publications, the paper should summarize what those new developments are. If both apply, then the paper should summarize how it supplements the previous work and what new development it adds.
>
> [@timokoch](https://github.com/timokoch) [@smiths](https://github.com/smiths) Do the suggestions above align with your comments?

This PR (https://github.com/ruohai0925/IAMReX/commit/8cb0f0b03f71525621e0368fba684c82abc51788) tries to include all points in the discussions.
