### 2025-06-28 ruohai0925/IAMReX#27 @ruohai0925
I will add more details of Figure 2.
