### 2025-06-30 ruohai0925/IAMReX#23 @timokoch
I believe this can be also addressed by complying with the REUSE specification. This was already started, however not completed. This specifies in particular that each file in the repo must have a license statement (either in the file, via a license file per file, or via a REUSE.toml file, see https://reuse.software/).

You can add your repo here https://api.reuse.software/register to automatically check for compliance with this standard.

In addition (to being fully REUSE compliant), you can also have a license file that GitHub detects but there the problem is that only one license is detected, so it might have to be a custom one that points to the two licenses and a statement of how they apply.

I believe it's also _necessary_ here to explain licensing in the README.
