+++ docs/_sources/Getting_Started.rst.txt
@@ -55,7 +55,10 @@ For example, if we want to compile in the ``FlowPastSphere``, refer to the follo
 
 #. Modify compilation parameters in GNUmakefile.
 
-    The compilation parameters depend on your computing platform. If you use MPI to run your program, then set USE_MPI = TRUE. If you are running the program with Nvidia GPU(CUDA), then set USE_CUDA = TRUE. When using GPU runtime, please make sure your CUDA environment is ok.
+    The compilation parameters depend on your computing platform. If you use MPI to run your program, then set ``USE_MPI = TRUE``. If you are running the program with Nvidia GPU(CUDA), then set ``USE_CUDA = TRUE``. When using GPU runtime, please make sure your CUDA environment is ok.
+
+    Here present an overview of typical MPI and CUDA configurations and environments : `OpenMPI <https://docs.open-mpi.org>`_, `MPICH <https://www.mpich.org/documentation/guides/>`_, `IntelMPI <https://www.intel.com/content/www/us/en/developer/tools/oneapi/mpi-library.html>`_, `CUDA <https://developer.nvidia.com/cuda-toolkit>`_.
+
 
 #. Compile
 
@@ -84,7 +87,35 @@ If you use ``MPI`` to run your program, you can type:
 
         mpirun -np how_many_processes amr3d.GNU.MPI.ex inputs.3d.flow_past_sphere
 
-This code typically generates subfolders in the current folder that are named ``plt00000``, ``plt00010``, etc, and ``chk00000``, ``chk00010``, etc. These are called plotfiles and checkpoint files. The plotfiles are used for visualization of derived fields; the checkpoint files are used for restarting the code.
+This code typically generates subfolders in the current folder that are named ``plt00000``, ``plt00010``, etc, and ``chk00000``, ``chk00010``, etc. These are called **plotfiles** and **checkpoint** files. The plotfiles are used for visualization of derived fields; the checkpoint files are used for restarting the code.
+
+
+Visualization
+--------------
+
+Various tools are available for data reading and processing , you can check `AMReX <https://erf.readthedocs.io/en/latest/Visualization.html>`_ for reference. As for the specific settings and output content, `IAMR <https://amrex-fluids.github.io/IAMR/RunningProblems.html#output-options>`_ provides more details.
+
+**IAMReX currently generates plotfile in the native AMReX format.** The particle data will be generated in the case directory as corresponding CSV files named according to the particle IDs, following the naming convention ``IB_Particle_<ID>.csv``
+
+.. figure:: ./GettingStarted/ParaView_Plotfiles.png
+    :align: center
+    :alt:  To visualize plotfiles, use ParaView to process and display the results.
+    :width: 5.00000in
+
+    To visualize plotfiles, use ParaView to process and display the results.
+
+Plotfiles can include the quantities of several simulation parameters as output(density, gradpx, gradpy, gradpz, x_velocity, y_velocity, z_velocity and tracer). We recommend using ``ParaView`` to process **plotfiles**.
+
+Checkpoints are used to resume simulations. You typically need to specify how often to save checkpoint files. To restart a simulation from a particular checkpoint, you must specify it in the ``restart`` parameters. These settings are usually defined under the ``amr`` field.
+
+::
+
+    # how many timesteps to save
+    amr.check_int = 4000
+    # which check point to restart
+    amr.restart = chk00010
+
+Furthermore, particle data processing scripts are provided herein `IBParticle2VTK <https://github.com/S-Explorer/IBParticle2VTK>`_. These utilities enable conversion of particle **CSV** files to **VTK** format, facilitating further analysis in ParaView.
 
 .. _sec:PhysicsParams:
 
[…]
