### 2025-07-01 #8080 @piyueh
This is just a note to track JOSS-related issues in the code repository and serve as a record of review activities:
* https://github.com/ruohai0925/IAMReX/issues/12
* https://github.com/ruohai0925/IAMReX/issues/13
* https://github.com/ruohai0925/IAMReX/issues/18
* https://github.com/ruohai0925/IAMReX/issues/21
* https://github.com/ruohai0925/IAMReX/issues/23
* https://github.com/ruohai0925/IAMReX/issues/24
* https://github.com/ruohai0925/IAMReX/issues/25
* https://github.com/ruohai0925/IAMReX/issues/26
* https://github.com/ruohai0925/IAMReX/issues/27
* https://github.com/ruohai0925/IAMReX/issues/28
* https://github.com/ruohai0925/IAMReX/issues/29
