+++ docs/_sources/Introduction_Chapter.rst.txt
@@ -19,7 +19,7 @@ Sample Cases
 ------------------
 
 In the first case, we test the order of convergence for in a 2D reversed single vortex problem. 
-A 2D circular drop with radius $R=0.15$ is placed at $(0.5,0.75)$ in a unit computational domain $[0, 1] \times [0, 1]$.
+A 2D circular drop with radius R=0.15 is placed at (0.5,0.75) in a unit computational domain [0, 1] x [0, 1].
 The velocity field is given by the stream function,
 
    .. math::
@@ -28,9 +28,9 @@ The velocity field is given by the stream function,
       \Psi(x, y, t)=\frac{1}{\pi} \sin ^{2}(\pi x) \sin ^{2}(\pi y) \cos \left(\frac{\pi t}{T}\right),
 
 in which the rotational period is $T=4.0$. The velocities in the $x$ and $y$ directions are defined as $u(x,y,t)=\partial \Psi / \partial y$ and $v(x,y,t)=-\partial \Psi / \partial x$, respectively.
-The periodic boundary condition is applied in both $x$ and $y$ directions. At $t=T$, the interfacial profile of the drop at the high resolution converges to its initial profile, which represents the exact solution. 
+The periodic boundary condition is applied in both x and y directions. At t=T, the interfacial profile of the drop at the high resolution converges to its initial profile, which represents the exact solution. 
 On the other hand, the circular shape of the drop is distorted for the simulation with low resolution.
-Here is the case path : `RSV <https://github.com/ruohai0925/IAMReX/tree/development/Tutorials/RSV>`_.
+Here is the case path: `RSV <https://github.com/ruohai0925/IAMReX/tree/development/Tutorials/RSV>`_.
 
 .. figure:: ./GettingStarted/RSV.jpeg
     :align: center
@@ -39,27 +39,32 @@ Here is the case path : `RSV <https://github.com/ruohai0925/IAMReX/tree/developm
 
     Profiles of drop interface in the Reverse Single Vortex (RSV) problem at t/T=1 after one rotation. Black line: Analytical Solution; Red line: 64*64; Blue line: 128*128; Green line: 256*256
 
-Case path : `RayleighTaylor <https://github.com/ruohai0925/IAMReX/tree/development/Tutorials/RayleighTaylor_LS>`_.
+The second Rayleigh--Taylor (RT) instability case is simulated here to validate the adaptive two-phase flow algorithms when small vorticity structures are involved.
+This instability phenomenon occurs for any perturbation to the interface between a lighter fluid ($\rho_2$) at the bottom and a heavier fluid ($\rho_1$) at the top.
+In the simulation, the computational domain is $[0,1] \times [0,4]$. The initial interface is given by $y(x)=2.0+0.1\cos(2\pi x)$.
+The density ratio is set to $\lambda=\rho_2/\rho_1=1/3$ and the Reynolds number is set to be $Re=\rho_1 g^{1/2}/\mu_1=3000$. Here is the case path: `RayleighTaylor <https://github.com/ruohai0925/IAMReX/tree/development/Tutorials/RayleighTaylor_LS>`_.
 
 .. figure:: ./GettingStarted/IAMR_LSAMR.png
     :align: center
-    :alt:  (a) Density profile at t/T=2.42 using Level Set (LS) method. (b) Density profile at t/T=2.42 using IAMReX convective scheme.
+    :alt:  (a) Density profile at t/T=2.42 using Level Set (LS) method with sharp interface. (b) Density profile at t/T=2.42 using IAMReX convective method with diffuse interface.
     :width: 5.00000in
 
-    Density profile: a. Density profile at t/T=2.42 using LS method. b. Density profile at t/T=2.42 using IAMReX convective scheme.
[…]
