### 2025-07-02 ruohai0925/IAMReX#25 @ruohai0925
We add an visualization session here: https://ruohai0925.github.io/IAMReX/Getting_Started.html#visualization. In this session, we point out the readers to refer to AMReX and IAMR for visulation using plotfiles.

In the key parameters session (https://ruohai0925.github.io/IAMReX/Getting_Started.html#key-parameters), we also add the parameters to guide the users to set up the checkpoint related parameters for restarting the simulation.
