### 2025-07-02 ruohai0925/IAMReX#26 @ruohai0925
All input scripts and raw postprocessing data of IAMReX have been uploaded into [Baidu Cloud](https://pan.baidu.com/s/1bZRoDunjBv7bqYL8CI3ASA?pwd=i5c2) for interested readers to reproduce the results. We add this information at the end of Sample Cases session.
