### 2025-07-02 ruohai0925/IAMReX#27 @ruohai0925
More explanation of different methods are added for Figure 2.
