### 2025-07-02 ruohai0925/IAMReX#28 @ruohai0925
The references have been added and a pointer to our previous work is given.
