### 2025-07-02 ruohai0925/IAMReX#29 @ruohai0925
As the previous issue (https://github.com/ruohai0925/IAMReX/issues/26), we added the case directory and explanations for the readers to reproduce our results. The raw data of Figure 4 is [here](https://pan.baidu.com/s/1bZRoDunjBv7bqYL8CI3ASA?pwd=i5c2#list/path=%2FResults%20Data%2FMonodisperse).
