### 2025-07-02 ruohai0925/IAMReX#23 @hswind4
@smiths Thank you for your suggestion. Would it be acceptable if we simply add a clarification about these two license in README?

In this repository, we’ve added REUSE-style license headers to each source file to explicitly indicate  the copyright and license information:
* For files inherited directly from IAMR, we retain the original IAMR license. If modified, there are explain in the license header.
https://github.com/ruohai0925/IAMReX/blob/3f3db9866100210cfea5fb569525ebebc844f0c9/Source/NavierStokes.cpp#L1-L5
* For files created from scratch, we apply a new license (BSD-3-Clause).
https://github.com/ruohai0925/IAMReX/blob/3f3db9866100210cfea5fb569525ebebc844f0c9/Source/Collision.H#L1-L5

I'm considering explain licensing in the README.  How do you think about that?
License display in sidebar, I'm trying to figure it out.
