### 2025-07-02 ruohai0925/IAMReX#23 @hswind4
Hi @timokoch,
Here is the REUSE compliance report: https://api.reuse.software/info/github.com/ruohai0925/IAMReX

We have essentially ensured that the source code, Test_IAMReX/, and Tutorial_profiling/ directories are properly licensed.

As for other parts of the repository — such as images, doc, rst, GitHub workflows, even LaTeX files  and Exec data— things become more complex. In some cases, like the Tutorials directory, we intentionally kept them similar to the original IAMR project for consistency.

So, would it be necessary to apply REUSE compliance to those files as well or is there a recommended level of compliance we should aim for?

Thank you for your patience, and I‘m looking forward to your further discussions.
