### 2025-07-02 ruohai0925/IAMReX#23 @timokoch
I think all files are equally part of the source code and for example for the tutorials it‘s also important to see whether they are from IAMR or newly added and which license applies.
It only follows the REUSE specifications if it‘s fully compliant. All parts should also be mentioned im the readme so it‘s clear what‘s forked and what‘s new.
