### 2025-07-03 ruohai0925/IAMReX#23 @hswind4
> [@hswind4](https://github.com/hswind4) Can you update README based on Timo's suggestions?

Got it.
