### 2025-07-05 ruohai0925/IAMReX#33 @timokoch
README states project is reuse compliant but its not
See https://api.reuse.software/info/github.com/ruohai0925/IAMReX
