### 2025-07-08 ruohai0925/IAMReX#33 @ruohai0925
@timokoch Now the project is reuse compliant with the fix of this PR: https://github.com/ruohai0925/IAMReX/pull/35.
