### 2025-07-11 ruohai0925/IAMReX#25 @smiths
I've gone through the changes and followed the instructions to make my own plots.  Everything went fine, so we can close this issue.  :smile:
