### 2025-07-12 ruohai0925/IAMReX#26 @smiths
@ruohai0925, the revisions to the sample cases look great.  I like that the input files are available in your Tutorials folder.  I don't think you need the files in the Baidu Cloud.  As long as people can reproduce the examples themselves, I think you are good.  (The Baidu Cloud also doesn't come across as professional as the rest of your work.)  :smile:
