### 2025-07-12 #8080 @smiths
@ruohai0925, except for sorting out the LICENSE issues (https://github.com/ruohai0925/IAMReX/issues/33, https://github.com/ruohai0925/IAMReX/issues/23), you have addressed my current feedback.  I'm still working through the rest of [my review](https://github.com/openjournals/joss-reviews/issues/8080#issuecomment-2815687576).  I should have time over the next two weeks to hopefully complete my review.
