### 2025-07-14 ruohai0925/IAMReX#23 @smiths
There is now a [REUSE.toml](https://github.com/ruohai0925/IAMReX/blob/development/REUSE.toml) file, but I would like @timokoch to sign off on this before we close this issue.
