### 2025-07-14 ruohai0925/IAMReX#23 @timokoch
I think the new REUSE.toml now gives a good overview what is licensed how and there is also an explanation now in the README about the licenses.

There is one strange wildcard use in REUSE.toml here: https://github.com/ruohai0925/IAMReX/blob/350d2d931073e5ff33c73b8b4f589d14694000a6/REUSE.toml#L95. Does this address all hidden files, i.e. files starting with "."? It looks to me like a wildcard referring to all files in the directory, but maybe in REUSE syntax this only refers to files starting with dot. In this case, all is fine.
