### 2025-07-14 ruohai0925/IAMReX#23 @hswind4
Hi Timokoch. In REUSE syntax, ".*" refers only to files starting with a dot. In our project, it includes .clang-tidy  .gitattributes  .gitignore and etc. The following is a simple test, if this line doesn't exist.
