### 2025-09-10 #8080 @piyueh
Hi @hklion, would you be interested in reviewing a submission for the Journal of Open Source Software? Your experience with AMReX and computational physics would provide exactly the perspective we need.
