### 2025-09-14 ruohai0925/IAMReX#48 @smiths
Not enough detail on particle collision detection
The IAMReX documentation is unclear on

- is particle collision in IAMR, or only in IAMReX?  In the [key differences section](https://ruohai0925.github.io/IAMReX/IAMRandIAMReX.html#key-differences-between-iamr-and-iamex), it sounds like IAMReX adds particles.  I believe that IAMR has particles (so they should be listed in the IAMR (Base)), but IAMReX adds particle collision.  Whatever the distinction, it should be made clear in this key differences section.
- how are particle collisions detected?  Spatial hashing is mentioned, but no details or references are provided.
- ACTM is mentioned, but no details are provided, or references given.
- What are the collision models?  An integer is used to represent the collision model in the [key parameters](https://ruohai0925.github.io/IAMReX/Getting_Started.html#key-parameters) section, but what are the available models?  (As an aside, wouldn't an enumerated type be better here than an integer?)
