### 2025-09-14 ruohai0925/IAMReX#50 @smiths
API Documentation?
I can find partial documentation in the key parameters section and through the examples, but where can I find the full API documentation?  I'm looking for this information so that I can check off the "Functionality documentation" in the IAMReX JOSS review.
