+++ docs/_sources/Results.rst.txt
@@ -161,7 +161,8 @@ We demonstrate the accuracy and efficacy of our codes for simulating clusters of
 
 Compared with them, our algorithm has a 72.5% grid reduction with :math:`d/h=16`. And it has a 62.5% Lagrangian markers reduction compared with the DLM method :cite:`sharma2022coupled,zeng2022subcycling` with :math:`d/h=16`.
 
-When the simulation reaches the steady state, the total pressure drop balances the IB force generated by all particles in the streamwise z direction. Following the equation in akiki et al., the theoretical drag force is given by
+When the simulation reaches the steady state, the total pressure drop balances the IB force generated by all particles in the streamwise z direction. 
+Following the equation in Akiki et al. :cite:`akiki2016immersed`, the theoretical drag force is given by
 
 .. math::
     F_{theory} = (\frac{\Delta p}{\Delta z}L_z)L_xL_y
