+++ docs/_sources/Results.rst.txt
@@ -19,7 +19,7 @@ The particle volume fraction (PVF), which joins the calculation of free motion u
 
     Sketch of :math:`\phi` at each corner of a grid cell with interface.
 
-The PVF variable is defined as the volume fraction of the fluid in the cell, which can be calculated by the level-set function :math:`\phi` 
+The PVF variable is defined as the volume fraction of the fluid in the cell, which can be calculated by the level-set function :math:`\phi`
 and the Heaviside function :math:`H`. Based on the level-set function :math:`\phi`, an approximation of PVF can be obtained by the following equation:
 
 .. math::
@@ -115,15 +115,17 @@ Flow Past Sphere
 
     The schematic of the flow passing through the spherical particles
 
-We validate the accuracy and efficacy of our adaptive solver by simulating a spherical particle in uniform flow with different particle Reynolds numbers :cite:`schiller1933uber,zhu2022particle`. The diameter of the particle is :math:`D_p = 1`, the computational domain is :math:`L_x \times L_y \times L_z = 20D_p \times 10D_p \times 10D_p`, the distance of the particle from the inlet is :math:`d = 5D_p` and located in the center of the yz plane. The inlet and outlet boundaries are applied in the x direction and the inlet velocity :math:`U` is :math:`1m/s`. Both y and z directions are periodic boundaries.
+We validate the accuracy and efficacy of our adaptive solver by simulating a spherical particle in uniform flow with different particle Reynolds numbers :cite:`schiller1933uber,zhu2022particle`. The diameter of the particle is :math:`D_p = 1`, the computational domain is :math:`L_x \times L_y \times L_z = 20D_p \times 10D_p \times 10D_p`, the distance of the particle from the inlet is :math:`d = 5D_p` and located in the center of the :math:`yz` plane. The inlet and outlet boundaries are applied in the :math:`x` direction and the inlet velocity :math:`U` is :math:`1m/s`. Both :math:`y` and :math:`z` directions are periodic boundaries.
 
-The influence of AMR on the simulation results is investigated by using the subcycling method with different levels. As shown in Fig.
+.. _figFPSR_ref:
 
 .. figure:: ./Results/FlowPastSphereResult.png
     :align: center
     :width: 5.0in
 
     The drag coefficient of the particle under uniform flows varies with Particle Reynolds number at different AMR levels.
+
+The influence of AMR on the simulation results is investigated by using the subcycling method with different levels. As shown in :numref:`figFPSR_ref`.
 
 Three different types of grid were selected: level 0 indicates the single-level grid containing no AMR, level 1 indicates the two-level grid, and level 2 indicates the three-level grid. For all three types of grid, the ratio of the diameter of the particles to the grid spacing on the finest level is 16. The drag force, including the contributions of Lagrangian force and the PVF function, is calculated by
 
@@ -135,14 +137,18 @@ The theoretical S-N law for calculating the drag coefficient of the shaped parti
 .. math::
     C_D = (24/Re_p)(1+0.15Re_p^{0.687}),
 
-which is proposed by Schiller :cite:`schiller1933uber`., and :math:`Re_p = UD_p/\nu` represents the particle Reynolds number. It can be seen from above Fig that the present results under different particle Reynolds numbers are in good agreement with S-N law. The fact that different levels of grid produce the nearly identical results validated the accuracy of our solver on the adaptive grid.
[…]
