### 2025-09-23 ruohai0925/IAMReX#48 @ruohai0925
For the first point, IAMR does not support resiolved particle collision. It only supports particle as passive scalars at this stage. I add more info to compare the particle support in the key differences section.

For the second point, we deleted the keyword "spatial hashing". We will add one section to explain the particle collision mode. That is one of missing parts in the documentation. Thanks for pointing it out!
