### 2025-10-15 #8080 @hklion
Understood on the upstream issues. Would you be able to add a note to the documentation that you have made additional changes to AMReX that only live in your fork? I had previously looked at that diff, but missed that there was more than just merge commits.

I've completed my initial review:

@smiths beat me to opening the issue about API documentation. Please let me know when that's updated as well.

I also opened an issue asking for pointers in the documentation for how to run a test suite, as well as about GPU test coverage.

As a heads up, my last day of work before maternity leave will be October 31 and I don't expect to fully return to work until early April. I will do my best to not delay acceptance once these two issues are addressed. I apologize that the review is coming up against my leave like this; my leave timeline has changed unexpectedly in the last few weeks.
