### 2025-11-03 #8080 @piyueh
> Hi [@piyueh](https://github.com/piyueh) , for Smiths's checklists, there are two parts waiting to be checked: Statement of need and Functionality documentation. The former was already included in the paper draft. The latter was fixed by these two commits: [ruohai0925/IAMReX#50](https://github.com/ruohai0925/IAMReX/issues/50) and [ruohai0925/IAMReX#48](https://github.com/ruohai0925/IAMReX/issues/48). Let us know whether we need to update some other parts.

Hi @smiths just a kind reminder to check in on the status of this review. The author has addressed your comments as described above. We appreciate your time and support!
