### 2025-11-13 #8080 @ruohai0925
Hi @piyueh, I fnished the additional tasks.

1. The release version is v1.0. Check it here: https://github.com/ruohai0925/IAMReX/releases/tag/v1.0

2. The zenodo page with proper license is created here: https://zenodo.org/records/17602905. Here is the DOI: https://doi.org/10.5281/zenodo.17602905

All other information is correct.
