### 2025-11-17 #8080 @ruohai0925
> [@ruohai0925](https://github.com/ruohai0925) Also, you may want to fix the case of the first name for this author on Zenodo: <img alt="Image" width="320" height="94" src="https://private-user-images.githubusercontent.com/7220966/515407669-f59c4c7e-0d62-4e6e-a8de-5defb4533a62.png?jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NjM0MTkyMDYsIm5iZiI6MTc2MzQxODkwNiwicGF0aCI6Ii83MjIwOTY2LzUxNTQwNzY2OS1mNTljNGM3ZS0wZDYyLTRlNmUtYThkZS01ZGVmYjQ1MzNhNjIucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MTExNyUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTExMTdUMjIzNTA2WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9MDc1ODFhYjdiMmU2ZTQ0OGIwNzc5NDcyNzIzZjc3MjJjYTYyODMwM2M3NTFlY2FhMTQ4ZTU2ZmFlZDFjMTQwYSZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.hoCgmPp6eU4Vo5L5l3VBIQvzBCHpytT4BuMojKDBI7k">

This is also updated.
