### 2025-11-18 #8080 @kyleniemeyer
@ruohai0925 I've made some minor edits to the paper in https://github.com/ruohai0925/IAMReX/pull/58, can you review and merge?
